better calculation for time to encode
