+++ packages/lambda/src/functions/helpers/concat-videos.ts
@@ -1,18 +1,10 @@
 import type {Codec, FfmpegExecutable} from '@remotion/renderer';
 import {combineVideos, RenderInternals} from '@remotion/renderer';
-import fs, {
-	createWriteStream,
-	existsSync,
-	mkdirSync,
-	promises,
-	rmdirSync,
-	rmSync,
-} from 'fs';
+import fs, {createWriteStream, promises} from 'fs';
 import path, {join} from 'path';
 import type {AwsRegion} from '../../pricing/aws-regions';
 import {
 	chunkKey,
-	CONCAT_FOLDER_TOKEN,
 	getErrorKeyPrefix,
 	REMOTION_CONCATED_TOKEN,
 	REMOTION_FILELIST_TOKEN,
@@ -65,7 +57,7 @@ const downloadS3File = async ({
 	});
 };
 
-const getAllFilesS3 = ({
+export const getAllFilesS3 = ({
 	bucket,
 	expectedFiles,
 	outdir,
@@ -180,52 +172,26 @@ const getAllFilesS3 = ({
 };
 
 export const concatVideosS3 = async ({
-	bucket,
-	expectedFiles,
 	onProgress,
 	numberOfFrames,
-	renderId,
-	region,
 	codec,
-	expectedBucketOwner,
 	fps,
 	numberOfGifLoops,
 	ffmpegExecutable,
 	remotionRoot,
-	onErrors,
+	files,
+	outdir,
 }: {
-	bucket: string;
-	expectedFiles: number;
 	onProgress: (frames: number) => void;
-	onErrors: (errors: EnhancedErrorInfo[]) => Promise<void>;
 	numberOfFrames: number;
-	renderId: string;
-	region: AwsRegion;
 	codec: LambdaCodec;
-	expectedBucketOwner: string;
 	fps: number;
 	numberOfGifLoops: number | null;
 	ffmpegExecutable: FfmpegExecutable;
 	remotionRoot: string;
+	files: string[];
+	outdir: string;
 }) => {
-	const outdir = join(RenderInternals.tmpDir(CONCAT_FOLDER_TOKEN), 'bucket');
-	if (existsSync(outdir)) {
-		(rmSync ?? rmdirSync)(outdir, {
-			recursive: true,
-		});
-	}
-
-	mkdirSync(outdir);
-	const files = await getAllFilesS3({
-		bucket,
-		expectedFiles,
-		outdir,
-		renderId,
-		region,
-		expectedBucketOwner,
-		onErrors,
-	});
-
 	const outfile = join(
 		RenderInternals.tmpDir(REMOTION_CONCATED_TOKEN),
 		'concat.' + RenderInternals.getFileExtensionFromCodec(codec, 'final')
